chore: rename slice
